[refactor] 패키지 수정
